+++ Output Files/BroadcastOutputs.h
@@ -1,0 +1,34 @@
+/*This file is part of Output Blaster.
+
+Output Blaster is free software : you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Output Blaster is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with Output Blaster.If not, see < https://www.gnu.org/licenses/>.*/
+
+#pragma once
+#include "Outputs.h"
+
+class CBroadcastOutputs : public COutputs
+{
+public:
+    CBroadcastOutputs(COutputs* a, COutputs* b);
+    virtual ~CBroadcastOutputs();
+
+    bool Initialize();
+    void Attached();
+
+protected:
+    void SendOutput(EOutputs output, UINT32 prevValue, UINT32 value);
+
+private:
+    COutputs* m_a;
+    COutputs* m_b;
+};