+++ win-game/simulate.py
@@ -4,7 +4,7 @@ import threading
 import random
 
 HOST = '127.0.0.1'
-PORT = 8000
+PORT = 37520
 FRAME = '\r\n'
 SEP = ' = '